Guns working
Got stuck on this for a few days. All gun mechanics are working now.
There was an issue with how the game handled inputs and I had to rework this.
Also I've still been cleaning up from the move, that is done now so I'll be able to work on this much more from now on.
If you are reading this expect significantly more commits from now on.

The game now has a gun, the gun can fire, it can be used and pick up from either hand, pressing the secondary button on the vr controller will eject the gun's magazin, the gun without a mag will accept a new one if it is close to it, the mag may be picked up, the controller properly disappears when a gun is picked up, the gun has limited ammo, ammo is tied to the magazin that the gun holds, firing the gun gives a sound effect and muzzle flash, the bullet properly leaves a hole, bullets leave a short tracer line